backend/requirements.txt
fastapi
uvicorn
python-dotenv

langchain
langchain-community
langchain-openai
langchain-text-splitters

faiss-cpu
pypdf

sqlalchemy
pyodbc
google-generativeai

sentence-transformers

